on-initialized .modems breaks example app
It seems there is an lazy initialization of ModemManager.modems
as consequence, the modems[0] are not available at the time
of start application and nothing works.

Similar OfonoConnMan contexts are not initialized. The binding must
check if array is defined and its length

I have added also the SimManager check to make sure that the Pin is
entered, because it prevents further testing.

Also window of example application wasn't showing